Move models to related packages
