+++ src/main/java/uk/gov/hmcts/reform/pbis/config/EmailTemplateMapping.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pbis.model;
+package uk.gov.hmcts.reform.pbis.config;
 
 public class EmailTemplateMapping {
