+++ src/main/java/uk/gov/hmcts/reform/pbis/notify/EmailToSend.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pbis.model;
+package uk.gov.hmcts.reform.pbis.notify;
 
 import java.util.Map;
