+++ src/test/java/uk/gov/hmcts/reform/pbis/notify/NotificationClientProviderTest.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.pbis.notify;
 
 import org.junit.Test;
 import uk.gov.hmcts.reform.pbis.ServiceNotFoundException;
-import uk.gov.hmcts.reform.pbis.model.EmailTemplateMapping;
+import uk.gov.hmcts.reform.pbis.config.EmailTemplateMapping;
 import uk.gov.service.notify.NotificationClient;
 import uk.gov.service.notify.NotificationClientApi;
